Refactoring: Memfault -> nRF Cloud, part 2
